orders/views.py
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework import status
from .models import Cart,Order
from .serializers import CartSerializer, OrderSerializer

