requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.6
django-cors-headers==3.10.0
django-environ==0.9.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.3.0
drf-friendly-errors==0.14
drf-friendly-errors-2==0.14.3
humanize==3.13.1
idna==3.3
mysqlclient==2.1.1
pillow==9.2.0
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.1
requests==2.27.1
sentry-sdk==1.5.10
sqlparse==0.4.1
typing_extensions==4.9.0
tzdata==2022.5
Unidecode==1.3.1
urllib3==1.26.8